Fixed gPTP path sequence generation
We only want to add our own identity to the head of the gPTP path
sequence if we are already the GM. The existing code checks
`compare_clock_identity_to_me` to see if we are the GM, which
returns 0 if true. If false this returns 1 half the time and -1
the other half, which causes an out-of-bounds array access failure
at runtime. Adding double negation stops this from happening.